# postman/specs/slangua-api.openapi.yaml
openapi: 3.0.3
info:
  title: SlangUA API
  version: 1.0.0
  description: >
    REST API для SlangUA — Telegram Mini App, що стилізує український текст у різні
    сучасні стилі мовлення.

    Ця специфікація згенерована на основі plans/docs/04-api.md, яка лишається
    джерелом правди для контрактів. Оновлюйте обидва документи синхронно зі змінами
    в src/routes/** та src/services/**.
servers:
  - url: /api/v1
    description: Версіонований базовий шлях (за проксі /api → backend)
...
paths:
  /auth/telegram:
    post:
      tags: [Auth]
      summary: Обмін Telegram initData на JWT access token
      description: >
        Перевіряє HMAC-SHA256 та TTL `auth_date`, створює/оновлює користувача,
        повертає access token у тілі й ставить refresh (`slangua_refresh`, HttpOnly)
        та читабельний `slangua_csrf` cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [initData]
              properties:
                initData:
                  type: string
                  description: Сирий рядок Telegram WebAppData
      responses:
        '200':
          description: Успішна автентифікація
          headers:
            Set-Cookie:
              description: Ставить slangua_refresh (HttpOnly) та slangua_csrf cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Ротація refresh-токена та видача нового access token
      description: >
        Refresh-токен передається лише в HttpOnly cookie `slangua_refresh`.
        Клієнт має надіслати заголовок `X-CSRF-Token`, що збігається з `slangua_csrf`.
      security:
        - refreshCookie: []
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: Токени зротовані
          headers:
            Set-Cookie:
              description: Оновлює slangua_refresh (HttpOnly) та slangua_csrf cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Інвалідизація поточного refresh-токена
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Сесію завершено
        '401':
          $ref: '#/components/responses/Error'
...
  /share/inline:
    post:
      tags: [Translate]
      summary: Створення короткоживучого токена для Telegram inline-share
      description: >
        Приймає рівно одне джерело — previewId або translationId (сирий текст ніколи
        не приймається). Потребує production-конфігурації inline-бота.
        Див. plans/docs/09-telegram-sharing.md.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [previewId]
                  properties:
                    previewId:
                      type: string
                      format: uuid
                - type: object
                  required: [translationId]
                  properties:
                    translationId:
                      type: integer
      responses:
        '200':
          description: Токен для inline-share
          content:
            application/json:
              schema:
                type: object
                properties:
                  inlineQuery:
                    type: string
                    example: s_<opaque-token>
                  expiresAt:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '410':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: slangua_refresh
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
      description: Значення з cookie slangua_csrf (double-submit CSRF)
...
  responses:
    Error:
      description: Стандартна помилка
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
    AccessTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
...